+++ packages/rocketh-deploy-proxy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # rocketh-deploy-proxy
 
+## 4.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.10.3
+  - rocketh-deploy@4.0.3
+
 ## 4.0.2
 
 ### Patch Changes