+++ packages/rocketh-export/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-export
 
+## 4.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.10.3
+
 ## 4.0.2
 
 ### Patch Changes