+++ packages/rocketh/CHANGELOG.md
@@ -1,5 +1,11 @@
 # rocketh
 
+## 0.10.3
+
+### Patch Changes
+
+- router allow extra abi
+
 ## 0.10.2
 
 ### Patch Changes